update_Ui

+++ Musicorum/Assets/Characters/RemovePlayerPreprefs.cs
@@ -4,10 +4,17 @@ using UnityEngine;
 
 public class RemovePlayerPreprefs : MonoBehaviour
 {
-
+    RemovePlayerPreprefs instance;
     // Use this for initialization
     private void Awake()
     {
+        if (instance != null && instance != this)
+        {
+            Destroy(this.gameObject);
+            return;//Avoid doing anything else
+        }
+        instance = this;
+        DontDestroyOnLoad(gameObject);
         PlayerPrefs.DeleteAll();
     }
 }
